added int cast on write call in new terminal printing code, and added an include of io.h on windows
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@6011 d7cf8633-e32d-0410-b094-e92efae38249

+++ sys_sdl.c
@@ -1,5 +1,6 @@
 
 #ifdef WIN32
+#include <io.h>
 #include "conio.h"
 #else
 #include <unistd.h>
@@ -56,7 +57,7 @@ void Sys_PrintToTerminal(const char *text)
 #endif
 	while(*text)
 	{
-		ssize_t written = write(1, text, strlen(text));
+		int written = (int)write(1, text, strlen(text));
 		if(written <= 0)
 			break; // sorry, I cannot do anything about this error - without an output
 		text += written;
